riscv128 : updating configuration files
Setting up the configuration files in such a way that the support for
building a riscv128 cross-environment tool can be bootstrapped.
Once the patch applied the configuration can be launched smoothly:
../configure --target=riscv128-unknown-elf --enable-maintainer-mode
Updated with upstream.

configure: regenerate rv128 compilation files

bfd: Addition of 128-bit support in elf file format and bfd

Supporting 128-bit architectures means having a dedicated elf128 file
format and then a bunch of somehow little modifications spread around
many files.
The elf128 data structures are very similar to the 64-bit ones, and use
the same trick to ensure alignement of the 128-bit integers: moving
fields around to ensure packed structs with as few dummy fields as
possible.
Still a few FIXME around to check, as certain values might well depend
on the data-size.

bfd/cpu-riscv.c: make rv128 the default for now

binutils/readelf.c: casts for proper compilation

Because the binutils guys removed the bfd_vma type and print function
during the summer, hackish solution to output 64 lower bits of vma on
warning and error messages.

fix: print size in readelf

rv128: Adding support in gas

The new instructions have to be known from gas, along with their
specificities. Part of this is close from the modifications done
in opcodes.

riscv128 : Add ld support

Making what's necessary to have a 128-bit riscv linker.
In essence:
  * changing Makefile.am
  * adding emulation scripts

+++ binutils/readelf.c
@@ -2723,7 +2723,7 @@ dump_relocations (Filedata *          filedata,
 		{
 		  putchar (' ');
 		  printf (_("<unknown addend: %" PRIx64 ">"),
-			  rels[i].r_addend);
+			  (unsigned long) rels[i].r_addend);
 		  res = false;
 		}
 	    }
@@ -2898,7 +2898,7 @@ dump_relocations (Filedata *          filedata,
 	  if (filedata->file_header.e_machine == EM_SPARCV9
 	      && rtype != NULL
 	      && streq (rtype, "R_SPARC_OLO10"))
-	    printf (" + %" PRIx64, ELF64_R_TYPE_DATA (inf));
+	    printf (" + %" PRIx64, (unsigned long)  ELF64_R_TYPE_DATA (inf));
 
 	  putchar ('\n');
 
@@ -7537,14 +7537,14 @@ process_program_headers (Filedata * filedata)
       else
 	printf (_("\nElf file type is %s\n"), get_file_type (filedata));
       printf (_("Entry point 0x%" PRIx64 "\n"),
-	      filedata->file_header.e_entry);
+	      (unsigned long) filedata->file_header.e_entry);
       printf (ngettext ("There is %d program header,"
 			" starting at offset %" PRIu64 "\n",
 			"There are %d program headers,"
 			" starting at offset %" PRIu64 "\n",
 			filedata->file_header.e_phnum),
-	      filedata->file_header.e_phnum,
-	      filedata->file_header.e_phoff);
+		filedata->file_header.e_phnum,
+		(unsigned long) filedata->file_header.e_phoff);
     }
 
   if (! get_program_headers (filedata))
@@ -8036,7 +8036,7 @@ get_32bit_elf_symbols (Filedata *filedata,
     {
       error (_("Section %s has an invalid sh_entsize of %#" PRIx64 "\n"),
 	     printable_section_name (filedata, section),
-	     section->sh_entsize);
+	     (unsigned long) section->sh_entsize);
       goto exit_point;
     }
 
@@ -8044,7 +8044,7 @@ get_32bit_elf_symbols (Filedata *filedata,
     {
       error (_("Section %s has an invalid sh_size of %#" PRIx64 "\n"),
 	     printable_section_name (filedata, section),
-	     section->sh_size);
+	     (unsigned long) section->sh_size);
       goto exit_point;
     }
 
@@ -8054,9 +8054,9 @@ get_32bit_elf_symbols (Filedata *filedata,
     {
       error (_("Size (%#" PRIx64 ") of section %s "
 	       "is not a multiple of its sh_entsize (%#" PRIx64 ")\n"),
-	     section->sh_size,
+	     (unsigned long) section->sh_size,
 	     printable_section_name (filedata, section),
-	     section->sh_entsize);
+	     (unsigned long) section->sh_entsize);
       goto exit_point;
     }
 
@@ -8089,8 +8089,8 @@ get_32bit_elf_symbols (Filedata *filedata,
 	{
 	  error (_("Index section %s has an sh_size of %#" PRIx64 " - expected %#" PRIx64 "\n"),
 		 printable_section_name (filedata, entry->hdr),
-		 entry->hdr->sh_size,
-		 section->sh_size);
+		 (unsigned long) entry->hdr->sh_size,
+		 (unsigned long) section->sh_size);
 	  goto exit_point;
 	}
     }
@@ -8153,7 +8153,7 @@ get_64bit_elf_symbols (Filedata *filedata,
     {
       error (_("Section %s has an invalid sh_entsize of %#" PRIx64 "\n"),
 	     printable_section_name (filedata, section),
-	     section->sh_entsize);
+	     (unsigned long) section->sh_entsize);
       goto exit_point;
     }
 
@@ -8161,7 +8161,7 @@ get_64bit_elf_symbols (Filedata *filedata,
     {
       error (_("Section %s has an invalid sh_size of %#" PRIx64 "\n"),
 	     printable_section_name (filedata, section),
-	     section->sh_size);
+	     (unsigned long) section->sh_size);
       goto exit_point;
     }
 
@@ -8171,9 +8171,9 @@ get_64bit_elf_symbols (Filedata *filedata,
     {
       error (_("Size (%#" PRIx64 ") of section %s "
 	       "is not a multiple of its sh_entsize (%#" PRIx64 ")\n"),
-	     section->sh_size,
+	     (unsigned long) section->sh_size,
 	     printable_section_name (filedata, section),
-	     section->sh_entsize);
+	     (unsigned long) section->sh_entsize);
       goto exit_point;
     }
 
@@ -8206,8 +8206,8 @@ get_64bit_elf_symbols (Filedata *filedata,
 	{
 	  error (_("Index section %s has an sh_size of %#" PRIx64 " - expected %#" PRIx64 "\n"),
 		 printable_section_name (filedata, entry->hdr),
-		 entry->hdr->sh_size,
-		 section->sh_size);
+		 (unsigned long) entry->hdr->sh_size,
+		 (unsigned long) section->sh_size);
 	  goto exit_point;
 	}
     }
@@ -8629,13 +8629,13 @@ process_section_headers (Filedata * filedata)
 	printf (_("In linked file '%s': "),
 		printable_string (filedata->file_name, 0));
       if (! filedata->is_separate || process_links)
-	printf (ngettext ("There is %d section header, "
+	printf (ngettext ("There is %ld section header, "
 			  "starting at offset %#" PRIx64 ":\n",
-			  "There are %d section headers, "
+			  "There are %ld section headers, "
 			  "starting at offset %#" PRIx64 ":\n",
 			  filedata->file_header.e_shnum),
-		filedata->file_header.e_shnum,
-		filedata->file_header.e_shoff);
+		(unsigned long) filedata->file_header.e_shnum,
+		(unsigned long) filedata->file_header.e_shoff);
     }
 
   if (!get_section_headers (filedata, false))
@@ -8715,9 +8715,9 @@ process_section_headers (Filedata * filedata)
       if (section->sh_entsize != expected_entsize)			\
 	{								\
 	  error (_("Section %d has invalid sh_entsize of %" PRIx64 "\n"), \
-		 i, section->sh_entsize);				\
+		 i, (unsigned long) section->sh_entsize);				\
 	  error (_("(Using the expected size of %" PRIx64 " for the rest of this dump)\n"), \
-		 expected_entsize);					\
+		 (unsigned long) expected_entsize);					\
 	  section->sh_entsize = expected_entsize;			\
 	}								\
     }									\
@@ -9583,8 +9583,8 @@ process_section_groups (Filedata * filedata)
 	      error (_("Section %s has sh_entsize (%#" PRIx64 ")"
 		       " which is larger than its size (%#" PRIx64 ")\n"),
 		     printable_section_name (filedata, section),
-		     section->sh_entsize,
-		     section->sh_size);
+		     (unsigned long) section->sh_entsize,
+		     (unsigned long) section->sh_size);
 	      continue;
 	    }
 
@@ -10691,8 +10691,8 @@ ia64_process_unwind (Filedata * filedata)
 	    printf ("'%s'", printable_section_name (filedata, unwsec));
 
 	  printf (_(" at offset %#" PRIx64 " contains %" PRIu64 " entries:\n"),
-		  unwsec->sh_offset,
-		  unwsec->sh_size / (3 * eh_addr_size));
+		  (unsigned long) unwsec->sh_offset,
+		  (unsigned long) unwsec->sh_size / (3 * eh_addr_size));
 
 	  if (slurp_ia64_unwind_table (filedata, & aux, unwsec)
 	      && aux.table_len > 0)
@@ -11266,7 +11266,7 @@ get_unwind_section_word (Filedata *                 filedata,
       if (rp->r_offset & 3)
 	{
 	  warn (_("Skipping unexpected relocation at offset %#" PRIx64 "\n"),
-		rp->r_offset);
+		(unsigned long) rp->r_offset);
 	  continue;
 	}
 
@@ -11297,7 +11297,7 @@ get_unwind_section_word (Filedata *                 filedata,
 	{
 	  error (_("Bad symbol index in unwind relocation "
 		   "(%" PRIu64 " > %" PRIu64 ")\n"),
-		 ELF32_R_SYM (rp->r_info), aux->nsyms);
+		 (unsigned long) ELF32_R_SYM (rp->r_info), (unsigned long) aux->nsyms);
 	  break;
 	}
 
@@ -12277,7 +12277,7 @@ dynamic_section_mips_val (Filedata * filedata, Elf_Internal_Dyn * entry)
 		get_dynamic_name (filedata, entry->d_un.d_val));
       else
 	printf (_("Interface Version: <corrupt: %" PRIx64 ">"),
-		entry->d_un.d_ptr);
+		(unsigned long) entry->d_un.d_ptr);
       break;
 
     case DT_MIPS_TIME_STAMP:
@@ -13760,7 +13760,7 @@ process_version_sections (Filedata * filedata)
 		      printable_section_name (filedata, section),
 		      section->sh_info);
 
-	    printf (_(" Addr: 0x%016" PRIx64), section->sh_addr);
+	    printf (_(" Addr: 0x%016" PRIx64), (unsigned long) section->sh_addr);
 	    printf (_("  Offset: 0x%08" PRIx64 "  Link: %u (%s)\n"),
 		    section->sh_offset, section->sh_link,
 		    printable_section_name_from_index (filedata, section->sh_link, NULL));
@@ -13906,7 +13906,7 @@ process_version_sections (Filedata * filedata)
 		      printable_section_name (filedata, section),
 		      section->sh_info);
 
-	    printf (_(" Addr: 0x%016" PRIx64), section->sh_addr);
+	    printf (_(" Addr: 0x%016" PRIx64), (unsigned long) section->sh_addr);
 	    printf (_("  Offset: 0x%08" PRIx64 "  Link: %u (%s)\n"),
 		    section->sh_offset, section->sh_link,
 		    printable_section_name_from_index (filedata, section->sh_link, NULL));
@@ -14071,9 +14071,9 @@ process_version_sections (Filedata * filedata)
 		      printable_section_name (filedata, section),
 		      total);
 
-	    printf (_(" Addr: 0x%016" PRIx64), section->sh_addr);
+	    printf (_(" Addr: 0x%016" PRIx64), (unsigned long) section->sh_addr);
 	    printf (_("  Offset: 0x%08" PRIx64 "  Link: %u (%s)\n"),
-		    section->sh_offset, section->sh_link,
+		    (unsigned long) section->sh_offset, section->sh_link,
 		    printable_section_name (filedata, link_section));
 
 	    off = offset_from_vma (filedata,
@@ -14928,9 +14928,9 @@ print_symbol (Filedata *           filedata,
       && psym->st_value > filedata->section_headers[psym->st_shndx].sh_size)
     warn (_("local symbol %s has a value (%#" PRIx64 ") which is larger than mergeable section %s's size (%#" PRIx64 ")\n"),
 	  strtab + psym->st_name,
-	  psym->st_value,
+	  (unsigned long) psym->st_value,
 	  printable_section_name_from_index (filedata, psym->st_shndx, NULL),
-	  filedata->section_headers[psym->st_shndx].sh_size);
+	  (unsigned long) filedata->section_headers[psym->st_shndx].sh_size);
 }
 
 static const char *
@@ -15015,7 +15015,7 @@ display_lto_symtab (Filedata *           filedata,
     {
       error (_("Section %s has an invalid sh_size of %#" PRIx64 "\n"),
 	     printable_section_name (filedata, section),
-	     section->sh_size);
+	     (unsigned long) section->sh_size);
       return false;
     }
 
@@ -15800,7 +15800,7 @@ target_specific_reloc_handling (Filedata *filedata,
 		if (leb_ret != 0 || reloc_size == 0 || reloc_size > 8)
 		  error (_("MSP430 ULEB128 field at %#" PRIx64
 			   " contains invalid ULEB128 value\n"),
-			 reloc->r_offset);
+			 (unsigned long) reloc->r_offset);
 		else if (sym_index >= num_syms)
 		  error (_("%s reloc contains invalid symbol index "
 			   "%" PRIu64 "\n"), "MSP430", sym_index);
@@ -15815,7 +15815,7 @@ target_specific_reloc_handling (Filedata *filedata,
 		      /* PR 21137 */
 		      error (_("MSP430 sym diff reloc contains invalid offset: "
 			       "%#" PRIx64 "\n"),
-			     reloc->r_offset);
+			     (unsigned long) reloc->r_offset);
 		  }
 
 		saved_sym = NULL;
@@ -15874,7 +15874,7 @@ target_specific_reloc_handling (Filedata *filedata,
 		    else
 		      error (_("MN10300 sym diff reloc contains invalid offset:"
 			       " %#" PRIx64 "\n"),
-			     reloc->r_offset);
+			     (unsigned long) reloc->r_offset);
 		  }
 
 		saved_sym = NULL;
@@ -15927,7 +15927,7 @@ target_specific_reloc_handling (Filedata *filedata,
 	    else
 	      error (_("RL78 sym diff reloc contains invalid offset: "
 		       "%#" PRIx64 "\n"),
-		     reloc->r_offset);
+		     (unsigned long) reloc->r_offset);
 	    value = 0;
 	    return true;
 
@@ -15937,7 +15937,7 @@ target_specific_reloc_handling (Filedata *filedata,
 	    else
 	      error (_("RL78 sym diff reloc contains invalid offset: "
 		       "%#" PRIx64 "\n"),
-		     reloc->r_offset);
+		     (unsigned long) reloc->r_offset);
 	    value = 0;
 	    return true;
 
@@ -16903,7 +16903,7 @@ apply_relocations (Filedata *filedata,
 	    {
 	      warn (_("skipping invalid relocation offset %#" PRIx64
 		      " in section %s\n"),
-		    rp->r_offset,
+		    (unsigned long) rp->r_offset,
 		    printable_section_name (filedata, section));
 	      continue;
 	    }
@@ -20323,7 +20323,7 @@ process_attributes_v2 (Filedata *filedata,
       if (op.err)
 	{
 	  error (_("Cannot parse subsection at offset %"PRIx64"\n"),
-	    sec_hdr->sh_size - remaining);
+	    (unsigned long) sec_hdr->sh_size - remaining);
 	  res = false;
 	  goto free_data;
 	}
